Create Timeline and Profile views
Create the structures to support a user's timeline and profile.
Create the layout for profile and the views/partials to fill out the
missing aspect of that profile layout (banner).
Convert profile views to use profile layout.
Make route, action, view, and tests for the Timeline (which also uses
the profile layout.)